refactor(architecture)!: reorganize domain and infrastructure layers
