+++ src/domain/constants/app.constants.ts
