+++ src/domain/constants/key.constant.ts
