+++ src/domain/entities/deep-partial.entity.ts
