+++ src/domain/entities/errors/strict-store.error.code.ts
