+++ src/domain/entities/errors/strict-store.error.ts
@@ -1,5 +1,5 @@
-import { STRICT_STORE_ERROR_CODE } from '@core/entities/errors/strict-store.error.code';
-import { STRICT_STORE_ERROR_MESSAGE } from '@core/entities/errors/strict-store.error.msg';
+import { STRICT_STORE_ERROR_CODE } from '@src/domain/entities/errors/strict-store.error.code';
+import { STRICT_STORE_ERROR_MESSAGE } from '@src/domain/entities/errors/strict-store.error.msg';
 
 type StrictStoreErrorCode = typeof STRICT_STORE_ERROR_CODE[keyof typeof STRICT_STORE_ERROR_CODE];
