+++ src/domain/entities/persistable.entity.ts
