+++ src/domain/entities/persistence-type.entity.ts
