+++ src/domain/types/phantom-type.symbol.ts
