+++ src/infrastructure/adapters/web-storage.adapter.ts
@@ -1,4 +1,4 @@
-import { KeyValueStoragePort } from '@core/ports/key-value-storage.port';
+import { KeyValueStoragePort } from '@src/domain/ports/key-value-storage.port';
 
 export const webStorageAdapter = (storage: Storage): KeyValueStoragePort => ({
   get: (key: string) => storage.getItem(key),
